Splitting NumString
