+++ Assets/Tayx/Graphy - Ultimate Stats Monitor/Scripts/Audio/AudioText.cs
@@ -12,6 +12,7 @@ using System.Collections;
 using System.Text;
 using UnityEngine.UI;
 using Tayx.Graphy.Utils;
+using Tayx.Graphy.Utils.NumString;
 
 namespace Tayx.Graphy.Audio
 {
