+++ Assets/Tayx/Graphy - Ultimate Stats Monitor/Scripts/Fps/FpsText.cs
@@ -13,6 +13,7 @@ using UnityEngine.UI;
 using System.Collections;
 using System.Text;
 using Tayx.Graphy.Utils;
+using Tayx.Graphy.Utils.NumString;
 
 namespace Tayx.Graphy.Fps
 {
